23_job

+++ db_controller.py
@@ -1,9 +1,9 @@
 # todo: check errors
-from sqlalchemy import create_engine, Column, Integer, String, ForeignKey, insert, select
+from sqlalchemy import create_engine, Column, Integer, String, ForeignKey, insert, select, delete
 from sqlalchemy.orm import Session, DeclarativeBase, relationship
 from sqlalchemy.orm import mapped_column
-import settings
 
+from settings import String
 
 class Base(DeclarativeBase):
     ...
@@ -37,6 +37,7 @@ class UserCart(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     uid = Column(ForeignKey('user.id', ondelete='CASCADE'), nullable=False, index=True)
     item = Column(ForeignKey('item.id', ondelete='CASCADE'), nullable=False, index=True)
+    count =Column(Integer)
 
 
 
@@ -44,7 +45,7 @@ class UserCart(Base):
 
 class Database():
     def __init__(self):
-        self.__engine = create_engine(settings.String.sqlite, echo=True)
+        self.__engine = create_engine(String.sqlite, echo=True)
         self.session = Session(self.__engine)
         Base.metadata.create_all(self.__engine)
     def insert(self,table,returning=None,**values):
@@ -56,9 +57,8 @@ class Database():
             self.session.execute(insert(table).values(**values))
             self.session.commit()
 
-
-    def update(self, table, where, values):
-        ...
+    def is_available_items(self,id_item,count):
+        return True if count<=self.select(Item.count,(Item.id==id_item,),one=True)[0] else False
     def update(self, table, filter: list, values: dict):
         self.session.query(table).filter(*filter).update(values)
         self.session.commit()
@@ -70,9 +70,16 @@ class Database():
             return self.session.query(table).filter(*filter).one() if one else self.session.query(table).filter(
                 filter).all()
 
-    def delete(self, table, filter: list, values: dict):
-        self.session.query(table).filter(*filter).delete()
-        self.session.commit()
+    def delete(self, table, filter: list,returning=None ):
+        if returning:
+            returns=self.session.execute(delete(table).returning(returning)).fetchone()
+            self.session.commit()
+            return returns
+        else:
+            self.session.query(table).filter(*filter).delete()
+            self.session.commit()
+
+
